Changed user_id in DB to Numeric

+++ src/user.py
@@ -1,5 +1,5 @@
 from db import Base, session, engine
-from sqlalchemy import Column, String, Integer, BigInteger
+from sqlalchemy import Column, String, Integer, Numeric
 
 
 # The user model: a user can have multiple GitHub usernames for different teams but one GitHub username for the same
@@ -9,7 +9,7 @@ class User(Base):
 
     # Fields
     unique_id = Column(Integer, primary_key=True)
-    user_id = Column(BigInteger)
+    user_id = Column(Numeric)
     user_team = Column(String)
     user_github = Column(String)
